+++ src/NBench/Tracing/NoOpBenchmarkTrace.cs
@@ -1,9 +1,13 @@
-﻿using System;
+﻿// Copyright (c) Petabridge <https://petabridge.com/>. All rights reserved.
+// Licensed under the Apache 2.0 license. See LICENSE file in the project root for full license information.
+
+using System;
 
 namespace NBench.Tracing
 {
+    /// <inheritdoc />
     /// <summary>
-    /// Default no-op implementation of <see cref="IBenchmarkTrace"/>. Does nothing.
+    /// Default no-op implementation of <see cref="T:NBench.IBenchmarkTrace" />. Does nothing.
     /// </summary>
     internal sealed class NoOpBenchmarkTrace : IBenchmarkTrace
     {
@@ -37,3 +41,4 @@ namespace NBench.Tracing
         }
     }
 }
+