+++ tests/NBench.Tests.Performance.WithDependencies/Actor/ActorThroughputSpec.cs
@@ -1,4 +1,7 @@
-﻿//-----------------------------------------------------------------------
+﻿// Copyright (c) Petabridge <https://petabridge.com/>. All rights reserved.
+// Licensed under the Apache 2.0 license. See LICENSE file in the project root for full license information.
+
+//-----------------------------------------------------------------------
 // <copyright file="ActorThroughputSpec.cs" company="Akka.NET Project">
 //     Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
 //     Copyright (C) 2013-2016 Akka.NET project <https://github.com/akkadotnet/akka.net>
@@ -248,3 +251,4 @@ namespace Akka.Tests.Performance.Actor
         }
     }
 }
+